attendance feature backend/requirements.txt
Flask
flask-cors
face_recognition
opencv-python
numpy
Pillow
SQLAlchemy
